refactor(depotcache): update tests to use depotcache
